Put HEFT code in its own package & clean up code
